server/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.30.0
greenlet==3.0.3
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP Client
httpx==0.28.0
aiohttp==3.9.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Monitoring & Logging
sentry-sdk[fastapi]==1.39.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
yfinance==0.2.36
pytz==2024.1

# Testing
pytest==8.3.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-httpx==0.35.0

# Development
black==23.12.1
ruff==0.1.9
mypy==1.8.0
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3